requirements.txt
numpy==1.24.4
opencv_python==4.10.0.84
pandas==2.0.3
Pillow==11.0.0
torch==2.4.1+cu121
torchvision==0.19.1+cu121
tqdm==4.67.0
wandb==0.19.1
